update
git-svn-id: http://svn.egroupware.org/egroupware/trunk@6710 923e3c2e-ae13-0410-ba18-ced7211e33de

+++ backup/templates/default/script_form.tpl
@@ -115,8 +115,11 @@
 				{
 					if ($archive[$i]['bdate'] <= $rdate)
 					{
-						unlink($dir . '/' . $archive[$i]['file']);
-						echo 'removed ' . $dir . '/' . $archive[$i]['file'] . "\n";
+						if ($archive[$i]['file'] && $archive[$i]['file'] != '')
+						{
+							unlink($dir . '/' . $archive[$i]['file']);
+							echo 'removed ' . $dir . '/' . $archive[$i]['file'] . "\n";
+						}
 					}
 				}
 			}
